Fix movie date reading
Previously was using EXIF[:create_date] to get the dates movies were
created. This was wrong for live photos and for videos in general,
possibly due in part to bad time zone parsing, or possibly due to
:create_date being the date the video was uploaded/saved. As a result,
videos were off by a few hours and didn't line up with photos.

Using EXIF[:creation_date] for videos instead seems to fix the problem.